+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add benchmarks for ECDSA-k256 and EdDSA-25519 ([#598](https://github.com/0xMiden/crypto/pull/598)).
 - Added `compute_challenge_k()` and `verify_with_unchecked_k()` methods to separate hashing and EC logic in EdDSA over Ed25519 ([#602](https://github.com/0xMiden/crypto/pull/602)).
 - Refactored `LargeSmt` to use flat `Vec<Word>` layout for in-memory nodes ([#591](https://github.com/0xMiden/crypto/pull/594)).
+- Fixed IES sealed box deserialization ([#616](https://github.com/0xMiden/crypto/pull/616)).
 
 ## 0.17.1 (2025-10-10)